Support conversion of Kupang-style OurWord files to USFM with REs.
- Adds new mechanism to allow .re files to output carriage return and
line feed in substitutions ($r and $n).
- Adds new mechanism to allow .re files to perform range-limited replacements
(:/start/end/find/substitute/ causes replacement of find with substitute,
but only between matches for start and the next following match for end.)
- Adds beginning of mechanism for processes not easily expressed with REs
to be written in C# as built-in processes. So far just one, footnotes.process.
- Adds actual RE scripts for the Kupang conversion, and a test suite
to confirm they work the same as the old ones.